introducing virtual deformation
problem: to test a solution you have to proceed deformation step after
deformation step. For each deformation step you know:
- which members will be deformed
- their current_deformable_length
BUT you don’t know:
- if there are connections preventing a deformation

So you know which members should be deformed, but not the amount of
deformation.

virtual deformation is my first try. it’s a sort of observer pattern.
I’ll add a document in the folder for further explanation