interface folder
In this folder I created you will find:
- isdh
- interface proposal

The first is a python module we could use as interface. I’ve explained
it to Daniel during today’s meeting.
You can find some information about the module in the interface
proposal folder.
It contains a .ppt which is the only thing you have to care about. The
rest is just a collection of images