New isdh
Hopefully the last version of the “Initial State - Deformation History”
module. Check out the examples and tell me if you want me to explain
something.
I came up with this idea together with Moustafa.
What has been changed from the previous version is that:
1. Daniel’s function will get 2 input arguments, the first one is the
list of components as before, the second one is a dictionary that
contains the deformation_histories.
2. the class Component now only stores info about the initial state (no
dictionary anymore).
3. a new class called DeformationHistory has been created. It contains
a dictionary that stores list of DeformationStep as before, but the
keys of the dictionary are now the Components not the solution_number
anymore.