Finish implementation of the gaps within the existing code
With this commit, I have finished the inclusion of gaps within the code
that was previously able just to solve the structure without gaps and
cross members. This code now has the ability to solve structures that
have loadpaths that don't start together from the same point and those
who have discontinuities within the loadpaths. One more thing that
hasn't been implemented (but can easily be done) which is the ability of
the code to solve structure that have gaps at the very end adjacent to
the firewall.