Submission of the first draft of the solver
This commit has the first draft of the solver. With this, the solver can
solve any structure and output OoDs into a text file. Still some things
are missing such as the interface, the breaking option for cross
components and the answer to the question I have raised about the
location of the firewall